A little more error handling in export.sh

+++ scrapy_espn/extract.sh
@@ -17,32 +17,34 @@ if [ "$1" != "" ]; then
 	} < "../data/team_list.csv"
 
 	if [ "$FOUND" == 1 ]; then
-		echo "scrapy crawl schedule -a team=$1 -a year=$2 -o $1_$2.csv -t csv"
-		scrapy crawl schedule -a team="$1" -a year="$2" -o "$1_$2_schedule".csv -t csv
+		if [ "$2" != "" ]; then
+			echo "scrapy crawl schedule -a team=$1 -a year=$2 -o $1_$2.csv -t csv"
+			scrapy crawl schedule -a team="$1" -a year="$2" -o "$1_$2_schedule".csv -t csv
 	
-		mv "$1_$2_schedule".csv ../data/
-		printf "\n\nLook for your teams schedule at ../data/$1_$2_schedule.csv\n\n"
+			mv "$1_$2_schedule".csv ../data/
+			printf "\n\nLook for your teams schedule at ../data/$1_$2_schedule.csv\n\n"
 
-		{
-			read
-			while IFS=, read game_id neutral_court
-			do
-				if [ "$neutral_court" == 1]; then
-					printf "\n\nSkipping game: $game_id since it is neutral court\n\n"
-				else
-					printf "\n\ncrawling game: $game_id\n\n"
-					scrapy crawl game -a game="$game_id" -o "$1_$2_games.csv" -t csv
-				fi
-			done 
-		} < "../data/$1_$2_schedule.csv"
+			{
+				read
+				while IFS=, read game_id neutral_court
+				do
+					if [ "$neutral_court" == 1]; then
+						printf "\n\nSkipping game: $game_id since it is neutral court\n\n"
+					else
+						printf "\n\ncrawling game: $game_id\n\n"
+						scrapy crawl game -a game="$game_id" -o "$1_$2_games.csv" -t csv
+					fi
+				done 
+			} < "../data/$1_$2_schedule.csv"
 
-		# Delete temporary schedule file
-		rm "../data/$1_$2_schedule.csv"
+			# Delete temporary schedule file
+			rm "../data/$1_$2_schedule.csv"
 
-		# Remove duplicate data lines (mostly headers)
-		cat -n "$1_$2_games.csv" | sort -uk2 | sort -nk1 | cut -f2- > "../data/$1_$2_games.csv"
-		rm "$1_$2_games.csv"
-
+			# Remove duplicate data lines (mostly headers)
+			cat -n "$1_$2_games.csv" | sort -uk2 | sort -nk1 | cut -f2- > "../data/$1_$2_games.csv"
+			rm "$1_$2_games.csv"
+		else
+			printf "\n\n Please provide a second argument containting the year you would like to crawl\n\n"
 	else
 		printf "\n\nCould not find a team that corresponds to team identifier: $1\n\n"
 	fi
